requirements.txt
openai>=1.40.0
httpx>=0.27.0
python-dotenv>=1.0.1
pymupdf>=1.24.0
pymilvus>=2.4.5
milvus-lite>=2.4.8
